requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
PyMuPDF
pandas
pydantic
python-dotenv